fix(journal): reset descriptor before truncating
If the node crashed after truncating, then the descriptor might have the old entry.
This could lead to incorrect position of the writer. To prevent that we reset
the position to 0 and flush it immediately before truncating the segment. Resetting
it to 0 is fine because the only consequence is that the on startup the write scan
the whole segment. Otherwise, the lastIndex and lastPosition in the descriptor is never
used.