fix: don't store deployment when created
When deployments were still distributed through deployment distribution,
we needed to store the deployment record so it could be re-distributed
later. However, we've switched distributing deployments over to the
generalized record distribution logic which no longer needs this record.

Since the record is no longer needed, it is also no longer cleaned up
later. So, we need to avoid storing this record. This requires a new
version of the Deployment Created event.

No state needs to be changed for this event, so we can use the no-op
event applier.